Fixing the checks of new robotPosition coordinates
It wants to compare to the length of each line's list within 'grid'
(ie., len(grid[0])) for the x position, and the overall number of
lists within 'grid' (ie., len(grid)) for the y position.

Also, since the indexes (in robotPosition) are 0-based, but the
lengths returned by len() are 1-based, the checks need to be
'>=' not '>', otherwise it would allow, for example, robotPosition
to be [6, 0], which would be off the end of the first list.